requirements.txt
python-dotenv
requests
Flask
flask_sqlalchemy
flask_login
flask_mail
SQLAlchemy
psycopg2
boto3
flask_mail